test/requirements.txt
# We can't use PIP dependencies for the production code due to portability with
# certain user environments, but it's fine if we use dependencies in tests.

beautifulsoup4==4.12.2
jsonschema==4.18.4
pyyaml==6.0.1
